+++ app/Http/Requests/Api/V1/ForgotPasswordRequest.php
@@ -1,12 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Api\V1;
 
-use App\Models\User;
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rule;
 
-class UpdateUserRequest extends FormRequest
+class ForgotPasswordRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -24,20 +22,10 @@ class UpdateUserRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'name' => [
-                'string',
-                'max:255',
-            ],
             'email' => [
-                'string',
+                'required',
                 'email',
-                'max:255',
-                Rule::unique(User::class),
-            ],
-            'password' => [
-                'string',
-                'regex:/^(?=.*?[a-zA-Z])(?=.*?[0-9])(=.*?[#?!@$%^&*-]){0,}.{8,}$/',
-            ],
+            ]
         ];
     }
 }
