+++ app/Http/Requests/Api/V1/UpdateUserRequest.php
@@ -1,10 +1,12 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Api\V1;
 
+use App\Models\Api\V1\User;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
 
-class ResetPasswordRequest extends FormRequest
+class UpdateUserRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -22,15 +24,17 @@ class ResetPasswordRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'email' => [
-                'required',
-                'email',
+            'name' => [
+                'string',
+                'max:255',
             ],
-            'token' => [
-                'required',
+            'email' => [
+                'string',
+                'email',
+                'max:255',
+                Rule::unique(User::class),
             ],
             'password' => [
-                'required',
                 'string',
                 'regex:/^(?=.*?[a-zA-Z])(?=.*?[0-9])(=.*?[#?!@$%^&*-]){0,}.{8,}$/',
             ],
