+++ app/Models/Api/V1/User.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace App\Models;
+namespace App\Models\Api\V1;
 
 // use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
@@ -43,4 +44,9 @@ class User extends Authenticatable
         'email_verified_at' => 'datetime',
         'password' => 'hashed',
     ];
+
+    public function reviews(): BelongsToMany
+    {
+        return $this->belongsToMany(Review::class);
+    }
 }
