+++ app/Notifications/PasswordResetNotification.php
@@ -2,12 +2,10 @@
 
 namespace App\Notifications;
 
-use App\Models\User;
+use App\Models\Api\V1\User;
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Mail\Mailables\Content;
 use Illuminate\Mail\Mailables\Envelope;
-use Illuminate\Notifications\Messages\MailMessage;
 use Illuminate\Notifications\Notification;
 use Illuminate\Queue\SerializesModels;
