+++ tests/Feature/Auth/AuthenticationTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature\Auth;
 
-use App\Models\User;
+use App\Models\Api\V1\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
